+++ CHANGELOG.md
@@ -9,3 +9,10 @@ Internal testing version
 - Errors in `use-utility-connect` now return a consistent `Error` object
 - `with-utility-connect` now matches the current implementation of `use-utility-connect`
   - **Breaking Change**: The `ready` prop is no longer passed down by the HOC. Please use `loading` instead.
+
+## 4.0.0
+
+- Renamed the package from `utility-connect-react` to `connect-react`. Please reinstall from npm.
+  - `useUtilityConnect` hook renamed to `useConnect`
+  - `withUtilityConnect` HoC renamed to `withConnect`
+  - `config.UtilityConnectToken` renamed to `config.ConnectToken`